Moves functions to export
